Move RenderHelpMarkdown in help_markdown.go
