+++ help.go
@@ -5,13 +5,8 @@ import (
 	"fmt"
 	"io"
 	"reflect"
-	"runtime"
 	"strings"
-	"syscall"
 	"unicode"
-	"unsafe"
-
-	"github.com/charmbracelet/glamour"
 )
 
 // HelpFormat represents the possible formats for baker help.
@@ -117,63 +112,6 @@ func PrintHelp(w io.Writer, name string, comp Components, format HelpFormat) err
 	}
 
 	return nil
-}
-
-// RenderHelpMarkdown calls PrintHelp withy format markdown but render the
-// markdown for terminal consulation.
-func RenderHelpMarkdown(w io.Writer, name string, comp Components) error {
-	r, _ := glamour.NewTermRenderer(
-		// detect background color and pick either the default dark or light theme
-		glamour.WithAutoStyle(),
-		// wrap output at specific width
-		glamour.WithWordWrap(int(terminalWidth())),
-	)
-
-	if err := PrintHelp(r, name, comp, HelpFormatMarkdown); err != nil {
-		return err
-	}
-
-	r.Close()
-	_, err := io.Copy(w, r)
-	return err
-}
-
-func terminalWidth() uint {
-	const (
-		maxWidth     = 140 // don't go over 140 chars anyway
-		defaultWidth = 110 // in case we can't get the terminal width
-	)
-
-	var w uint
-
-	defer func() {
-		if err := recover(); err != nil {
-			w = defaultWidth
-		}
-	}()
-
-	if runtime.GOOS == "windows" {
-		// On windows assume 120 character wide terminal since the subsequent
-		// method only works on nix systems.
-		return 120
-	}
-
-	ws := &struct{ Row, Col, Xpixel, Ypixel uint16 }{}
-	retCode, _, errno := syscall.Syscall(syscall.SYS_IOCTL,
-		uintptr(syscall.Stdout),
-		uintptr(syscall.TIOCGWINSZ),
-		uintptr(unsafe.Pointer(ws)))
-
-	if int(retCode) == -1 {
-		panic(errno)
-	}
-
-	if ws.Col > maxWidth {
-		return maxWidth
-	}
-
-	w = uint(ws.Col)
-	return w
 }
 
 type baseDoc struct {
